frontend + backend
frontend, backend conectados y funcionando hasta crear auto

+++ backend_tingeso/src/main/java/com/example/backend_tingeso/controllers/RepairController.java
@@ -22,8 +22,8 @@ public class RepairController {
     RepairService repairService;
 
     @GetMapping("/patent/{patent}")
-    public float queso(@PathVariable String patent) {
-        float a = repairService.getCost(patent);
+    public double queso(@PathVariable String patent) {
+        double a = repairService.getCost(patent);
         return a;
     }
 }
